Factor out method to retrieve an ElasticClient
